unbounded: add unbounded endpoint
The unbounded endpoint can be configured (role) to run as an inbound (peer),
and outbound (client), or as both at the same time.

When running as inbound, it builds a WebRTC peer connection with another
unbounded endpoint running as client role, receiving TCP/UDP requests
over QUIC and routes to singbox.

When running as outbound, it builds a WebRTC peer connection with
another unbounded endpoint running as peer role, receiving TCP/UDP
requests from singbox and hands over through the unbouned QUIC layer to
the peer.

For TCP it uses QUIC streams and for UDP it uses QUIC datagram.